# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UnBusy OpenApi Documentation",
        "description": "UnBusy is a small public transport application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "UnBusy API Server"
        }
    ],
    "paths": {
...
        "/buses": {
            "get": {
                "tags": [
                    "Bus"
                ],
                "summary": "Get a list of all buses",
                "operationId": "bb4526731b4cbb1a8e10a02da4c2412a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Bus"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bus"
                ],
                "summary": "Create a new bus",
                "operationId": "df1fab631f549ed09727f347554c2f74",
                "requestBody": {
                    "description": "Bus object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bus created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bus"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/buses/{bus}": {
            "delete": {
                "tags": [
                    "Bus"
                ],
                "summary": "Delete a bus",
                "operationId": "1363bde87f34438284d8a049e34a992f",
                "parameters": [
                    {
                        "name": "bus",
                        "in": "path",
                        "description": "ID of the bus that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bus/update-location": {
            "post": {
                "tags": [
                    "Bus"
                ],
                "summary": "Updates the current location of a bus",
                "description": "Updates the current location of a bus",
                "operationId": "updateBusLocation",
                "requestBody": {
                    "description": "The key, longitude and latitude of the update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "longitude",
                                    "latitude"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": 123
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -122.4194
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.7749
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Bus line location updated successfully, busLocationUpdateEvent dispatched"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bus-lines": {
            "get": {
                "tags": [
                    "Bus Lines"
                ],
                "summary": "Get collection of bus lines",
                "description": "Returns a collection of bus lines",
                "operationId": "getBusLinesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusLineCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bus Lines"
                ],
                "summary": "Store new bus line",
                "description": "Returns the newly created bus line",
                "operationId": "storeBusLine",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusLineCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusLine"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bus-lines/{busLine}": {
            "get": {
                "tags": [
                    "Bus Lines"
                ],
                "summary": "Get bus line by ID",
                "description": "Returns a single bus line object identified by its ID",
                "operationId": "getBusLineById",
                "parameters": [
                    {
                        "name": "busLine",
                        "in": "path",
                        "description": "ID of the bus line to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusLine"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus line not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bus Lines"
                ],
                "summary": "Delete a bus line",
                "description": "Delete a specific bus line by its ID",
                "operationId": "deleteBusLine",
                "parameters": [
                    {
                        "name": "busLine",
                        "in": "path",
                        "description": "ID of the bus line to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bus line deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bus line not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BusLineCollection": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "FeatureCollection"
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/BusLineFeature"
                                },
                                {
                                    "$ref": "#/components/schemas/BusStopFeature"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "example": [
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Stari aerodrom - Donja gorica",
                            "type": "bus-line",
                            "id": 29
                        },
                        "geometry": {
                            "type": "LineString",
                            "coordinates": [
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ]
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Tržni centar (A)",
                            "id": 336,
                            "type": "bus-stop",
                            "line_id": 29
                        },
                        "geometry": {
                            "type": "Point",
                            "coordinates": [
                                42.43146669,
                                19.27915313
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Donja Gorica (A)",
                            "id": 337,
                            "type": "bus-stop",
                            "line_id": 29
                        },
                        "geometry": {
                            "type": "Point",
                            "coordinates": [
                                42.43133438,
                                19.28015506
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Donja gorica - Rogami ",
                            "type": "bus-line",
                            "id": 30
                        },
                        "geometry": {
                            "type": "LineString",
                            "coordinates": [
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ]
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Donja Gorica (A)",
                            "id": 337,
                            "type": "bus-stop",
                            "line_id": 30
                        },
                        "geometry": {
                            "type": "Point",
                            "coordinates": [
                                42.43133438,
                                19.28015506
                            ]
                        }
                    }
                ]
            },
...
            "BusLineFeature": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Feature"
                    },
                    "properties": {
                        "type": "object",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/BusLineProperties"
                            }
                        ]
                    },
                    "geometry": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "LineString"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": 19.27913681
                                    }
                                },
                                "example": [
                                    [
                                        19.27913681,
                                        42.43142816
                                    ],
                                    [
                                        19.33333333,
                                        42.33333333
                                    ]
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "BusLineProperties": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Stari aerodrom - Donja gorica"
                    },
                    "type": {
                        "type": "string",
                        "example": "bus-line"
                    },
                    "id": {
                        "type": "integer",
                        "example": 29
                    }
                },
                "type": "object"
            },
...
            "BusCreate": {
                "properties": {
                    "bus_line_id": {
                        "type": "integer",
                        "example": "122"
                    }
                },
                "type": "object",
                "example": {
                    "bus_line_id": 122
                }
            },
            "BusLineCreate": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "FeatureCollection"
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/BusLineCreateFeature"
                                },
                                {
                                    "$ref": "#/components/schemas/BusStopCreateFeature"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "example": [
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Stari aerodrom - Donja gorica",
                            "type": "bus-line"
                        },
                        "geometry": {
                            "type": "LineString",
                            "coordinates": [
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ],
                                [
                                    19.27913681,
                                    42.43142816
                                ]
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Tržni centar (A)",
                            "type": "bus-stop"
                        },
                        "geometry": {
                            "type": "Point",
                            "coordinates": [
                                42.43146669,
                                19.27915313
                            ]
                        }
                    },
                    {
                        "type": "Feature",
                        "properties": {
                            "name": "Donja Gorica (A)",
                            "type": "bus-stop"
                        },
                        "geometry": {
                            "type": "Point",
                            "coordinates": [
                                42.43133438,
                                19.28015506
                            ]
                        }
                    }
                ]
            },
            "UnauthenticatedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The type field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The type field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Bus": {
                "title": "Bus",
                "description": "Bus model",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the bus.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "key": {
                        "description": "The key of the bus.",
                        "type": "string"
                    },
                    "bus_line_id": {
                        "description": "The ID of the bus line that this bus belongs to.",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}